Separate two test cases. Separate JS code from `stock-ticker` template.
Two test cases that are possible are `stock-ticker` and
`stock-ticker-raw`. As we separated JS code from `stock-ticker`
template, and also the name of class for both `stock-ticker` and
`stock-ticker-raw` are the same, thus this will introduce error of
duplicate class name usage.

Thus we separate them into two test cases with link to click at the
bottom to test each one.