Rename stockticker (raw) as StockTickerRaw. Add test suites.
There's a problem when testing with Jasmine as it's duplicate name of
StockTicker. Thus we resolve this by renaming it to be different to each
other.

Added Jasmine test cases for both of element.
Testing with Polymer needs a little of effort. In this case, we involve
karma-polymer on NPM to do the work. Notice karma.conf.js file for more
info on configuration.